remove ActionBarTransaction
