+++ src/shared/components/ActionBar/index.ts
@@ -1,2 +1,1 @@
 export * from './ActionBar'
-export * from './ActionBarTransaction'
